+++ DoAnCuoiKy/DoAnCuoiKy/Model/Entities/InformationLibrary/Kho/BookImportTransaction.cs
@@ -8,9 +8,8 @@ namespace DoAnCuoiKy.Model.Entities.InformationLibrary.Kho
         public Guid? BookId { get; set; }
         public Book? book { get; set; }
         public int Quantity { get; set; }  // Số lượng nhập
-
-        public float UnitPrice { get; set; }
-        public float TotalPrice { get; set; }
+        public decimal UnitPrice { get; set; }
+        public decimal TotalPrice { get; set; }
         public TransactionType TransactionType { get; set; }
     }
 }
